system and APIC .h/.cpp: type mistakes
main.cpp: removing deadspace

LAPIC.cpp/.h: LAPIC can be written to, debugging tools are not helpful in this case, since it is not real memory, the memory mapping of the register always shows 0.
These registers must be written to using a full write of 32 bits and, specifically the SPV, must not affect lowest 4 bits.

IDT.s/.h/.cpp Specifying the number of IDT stubs in one place.

type mistake