It currently does not make sense to not force enabling SIMD nor to check if the simd is enabled when doing a simd_copy because SIMD requires the compiler to have it's SIMD flag set to enable SIMD instructions. In order to have my OS check for SIMD support and only use it if enabled, I have to write my own SIMD library supporting up to my highest desired SIMD mode and to use that without setting the -msse2 compiler flag.
For now, I have force defined SIMD as enabled but this should either be a compiler flag or the approach should be what I specified above where the OS determines whether the CPU supports SIMD (and to what level).

It may be enough to edit the existing emmintrin.h from the compiler library just enough to support a memcpy by copying the source code and bodging it to avoid using flag specific attribute tags.