Scheduling work started. First successful task switch. Need to add scheduling functionality (look up next task) and timekeeping and so on. Need to add exit functionality also (using aligned_free).
RTC, Scheduler, Shell, Term: renaming "instance" pointers for singletons to be type descriptive.

Memory: Adding aligned malloc (and free) finally.

Process: converted process into a struct for easier handling of data. Added priority and an initialiser with default vals.

Shell: added logging. Made Shell a singleton, which I may not want in the end. This is because I made Shell the default main loop. This is probably not permanent and I may figure out a different way to do this later. This doesn't matter for as long as there is only one Terminal as well.

CPU.h: trimmed cpu_context_t

GDT: moved the segment definitions to the header so that they can be accessed by other things.

IDT: include GDT for segment definition access.

LAPIC: removed handling -> moved to the scheduler files.