async_io queue working!
Right so turns out, if you store process.context->eax as a pointer to store the result in, and then write that before storing the context, it gets overwritten by the version of registers in the stack. therefore the value of SYSCALL_t::READ is stored, loaded and returned instead of n_read.