Adding asynchronous read functionality (#6)
* async_io: Adding asynchronous read functionality

The async_read is only used via syscalls because it doesn't make sense to call async read during kernel init.

ASYNC_READ compile option added to cmake because it's slower than sync read (but better in multithreaded situations of course!)

Added context_switch_period_us setting to compile times: 1ms for debug and 100us for release.

Signed-off-by: Artie Poole <stuart.poole@canonical.com>

* scheduler: elapsed ms usually < 1 so we should be using a floating point number.

* aysnc_io: add half-assed concept of busy device

* bloody mismatched linter

* async_io queue working!

Right so turns out, if you store process.context->eax as a pointer to store the result in, and then write that before storing the context, it gets overwritten by the version of registers in the stack. therefore the value of SYSCALL_t::READ is stored, loaded and returned instead of n_read.

* workflow changes and linter

---------

Signed-off-by: Artie Poole <stuart.poole@canonical.com>